requirements.txt
rich==13.7.1
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
fastapi==0.110.2
h11==0.14.0
idna==3.7
pydantic==2.7.1
pydantic_core==2.18.2
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
python-dotenv==1.0.1
ibm-watson-machine-learning==1.0.356
